pyproject.toml
[project]
name = "predictions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "seaborn>=0.13.0",
    "alpaca-py>=0.43.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]
